Adição de mecânicas
- Adição de função que verifica clicks dentro de um objeto;

- Adição de função que monta uma struct "objeto" a partir dos dados enviados a ela;

- Adição do sprite do campo de batalha e do barco, ambos provisórios;

- Adição da primeira troca de fases até o momento, indo para o teste de bossFight;

- Adição da struct do barco, utilizada nas lutas contra chefes;

- Mudanças de finalização nas telas, ambas sendo encerradas com uma variável própria, o que antes era feito com uma variável única para todas as telas.

+++ Mecanicas/verificadorDeClick/verificadorDeClick.c
@@ -9,7 +9,7 @@
 
 #include "../../Structs/objeto.h"
 
-bool verificadorDeClick(float clickX, float clickY, Objeto* objeto) {
+bool verificadorDeClick(int clickX, int clickY, Objeto* objeto) {
 
     if ( 
         clickX >= objeto->posicaoX && clickX <= objeto->posicaoX + objeto->largura &&
